+++ lib/utils/extensions/l10n_extension.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/widgets.dart';
+
+import 'package:wevacalc/utils/l10n/app_localizations.dart';
+
+/// Extension para acesso simplificado ao AppLocalizations via context.
+extension L10nExtension on BuildContext {
+  AppLocalizations get l10n => AppLocalizations.of(this)!;
+}