+++ lib/utils/l10n/app_localizations_en.dart
@@ -1,0 +1,22 @@
+// ignore: unused_import
+import 'package:intl/intl.dart' as intl;
+import 'app_localizations.dart';
+
+// ignore_for_file: type=lint
+
+/// The translations for English (`en`).
+class AppLocalizationsEn extends AppLocalizations {
+  AppLocalizationsEn([String locale = 'en']) : super(locale);
+
+  @override
+  String get appTitle => 'WevaCalc';
+
+  @override
+  String get calculator => 'Calculator';
+
+  @override
+  String get history => 'History';
+
+  @override
+  String get settings => 'Settings';
+}