+++ lib/utils/l10n/app_localizations_es.dart
@@ -1,0 +1,22 @@
+// ignore: unused_import
+import 'package:intl/intl.dart' as intl;
+import 'app_localizations.dart';
+
+// ignore_for_file: type=lint
+
+/// The translations for Spanish Castilian (`es`).
+class AppLocalizationsEs extends AppLocalizations {
+  AppLocalizationsEs([String locale = 'es']) : super(locale);
+
+  @override
+  String get appTitle => 'WevaCalc';
+
+  @override
+  String get calculator => 'Calculadora';
+
+  @override
+  String get history => 'Historial';
+
+  @override
+  String get settings => 'Configuración';
+}